[non-related] chore(i18n): Initialize i18n directory structure
For MSPCM's plan of going global, our documentation needs i18n support as well, and
here we eventually confirmed the directory structure of our documentation:

First, the `docs` directory is the root directory of the whole documentation project, bearing English
as the default locale. The directories (except directories of other locales) under `docs` should be
presented in English.

Next, in `docs/{language-code}` directories are the locale directories for various languages, which keep
the same structure as the root directory.

Also, the asset files of different locales are put sepretedly in the `assets` folder under its locale directory.

Signed-off-by: crrashh1542 <kotsuki@crrashh.com>