[non-related] refactor(ci): Rewrite the CI logic
As was said in #14, while performing CI for PRs, the workflow will run into an
error caused by being unable to push built distributions to GitHub Pages. So
this commit adds a workflow configuration file for PRs, which has Actions
build production pages only, instead of pushing to GitHub Pages.

Also, the whole Workflow structure are completely refactored, because most
of the content are ported from https://github.com/crrashh1542/win-up-to-date.

Signed-off-by: crrashh1542 <kotsuki@crrashh.com>